restructured files and directories
