+++ group_defender/__init__.py
@@ -1,5 +1,4 @@
 from group_defender.constants import *
 from group_defender.feedback import feedback_cov_handler
 from group_defender.payment import send_payment_options
-from group_defender.file import process_file
-from group_defender.scan_url import check_url
+from group_defender.defend import *
