+++ group_defender/defend/photo.py
